Library Management System Application
Created a fully functional library management system where books can be added, removed and displayed along with modifying staff members. Also there is a login page with ability to change username and password. This project was coded with exploitation of MySQL, javaswing and ORM framework(mysql connector) to connect between the database and the GUI.